CAMEL-14020 - Migrate getOut to getMessage in camel-jms. Use lambda where it's possible.

+++ components/camel-jms/src/test/java/org/apache/camel/component/jms/issues/JmsResequencerTest.java
@@ -47,11 +47,11 @@ public class JmsResequencerTest extends CamelSpringTestSupport {
         result.expectedMessageCount(100);
 
         for (int i = 0; i < 100; i++) {
-            result.message(i).body().isEqualTo(Integer.valueOf(i + 1));
+            result.message(i).body().isEqualTo(i + 1);
         }
 
         for (int i = 100; i > 0; i--) {
-            template.sendBodyAndHeader(endpoint, Integer.valueOf(i), "num", Long.valueOf(i));
+            template.sendBodyAndHeader(endpoint, i, "num", (long) i);
         }
 
         assertMockEndpointsSatisfied();
